# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "World Pool API",
        "description": "API para gerenciamento de bolões da Copa do Mundo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Lista todas as equipes",
                "operationId": "fc1c5dd9ef530faf6e8bba930cad296c",
                "responses": {
                    "200": {
                        "description": "Lista de equipes"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/teams/{id}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Retorna detalhes de uma equipe específica",
                "operationId": "ab05003decb5cbc16154382836f51e3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipe encontrada"
                    },
                    "404": {
                        "description": "Equipe não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/groups/{id}/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Lista todas as equipes de um grupo",
                "operationId": "1db044a7dfd347ccea52866d7f47701f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipes do grupo"
                    },
                    "404": {
                        "description": "Grupo ou equipes não encontradas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token obtido via POST /api/login",
                "scheme": "bearer"
            }
        }
    },
...
}